requirements.txt
autopep8==1.7.0
flake8==5.0.4
mccabe==0.7.0
numpy==1.23.2
opencv-python==4.6.0.66
pandas==1.4.3
Pillow==9.2.0
pycodestyle==2.9.1
pyflakes==2.5.0
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
toml==0.10.2
tqdm==4.64.0
